feature(#56): At initial connection don't fail in a loop
 but allow to change the settings and save to try again
